FIX : 컴포넌트, 컨텍스트 파일정리 심화완료
