+++ packages/assignment/src/context/NotificationContext.tsx
@@ -1,0 +1,66 @@
+import React, {
+  useState,
+  createContext,
+  useContext,
+  PropsWithChildren,
+} from 'react'
+import { useCallback, useMemo } from '../@lib/hooks'
+
+interface Notification {
+  id: number
+  message: string
+  type: 'info' | 'success' | 'warning' | 'error'
+}
+
+interface NotificationContextType {
+  notifications: Notification[]
+  addNotification: (message: string, type: Notification['type']) => void
+  removeNotification: (id: number) => void
+}
+
+const NotificationContext = createContext<NotificationContextType | null>(null)
+
+export const useNotification = () => {
+  const context = useContext(NotificationContext)
+  if (context === null) {
+    throw new Error(
+      'useNotification must be used within a NotificationProvider'
+    )
+  }
+  return context
+}
+
+export const NotificationProvider: React.FC<PropsWithChildren> = ({
+  children,
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([])
+
+  const addNotification = useCallback(
+    (message: string, type: Notification['type']) => {
+      const newNotification: Notification = {
+        id: Date.now(),
+        message,
+        type,
+      }
+      setNotifications((prev) => [...prev, newNotification])
+    },
+    []
+  )
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id)
+    )
+  }, [])
+
+  const value = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification]
+  )
+
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  )
+}
