+++ packages/assignment/src/context/ThemeContext.tsx
@@ -1,0 +1,35 @@
+import React, {
+  useState,
+  createContext,
+  PropsWithChildren,
+  useContext,
+} from 'react'
+import { useCallback, useMemo } from '../@lib/hooks'
+
+type Theme = 'light' | 'dark'
+
+interface ThemeContextType {
+  theme: Theme
+  toggleTheme: () => void
+}
+
+export const ThemeContext = createContext<ThemeContextType | null>(null)
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext)
+  if (context === null) {
+    throw new Error('useTheme must be used within a ThemeProvider')
+  }
+  return context
+}
+
+export const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [theme, setTheme] = useState<Theme>('light')
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === 'light' ? 'dark' : 'light'))
+  }, [])
+
+  const value = useMemo(() => ({ theme, toggleTheme }), [theme, toggleTheme])
+
+  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
+}
